requirements.txt
# Launcher / GUI entry (run from flez-bot root: python launcher.py).
# Full bot dependencies: see bot_runelite_IL/requirements.txt
PySide6>=6.5.0
python-dotenv>=1.0.1
